refactor: restructure actions folder
