+++ src/components/ContactDetail.tsx
@@ -11,7 +11,7 @@ import { useContacts } from '@/app/ContactsProvider'
 import ContactDetailPhone from './ContactDetailPhone'
 import phone from 'phone'
 import { useFormState } from 'react-dom'
-import { saveContact, State } from '@/actions/saveContact'
+import { saveContact, State } from '@/actions/common/saveContact'
 import LoadingSpinner from './common/spinner'
 import { useNotification } from '@/app/NotificationsProvider'
 import { Session } from 'next-auth'
